Move statistics charts
