+++ frontend/components/statistics/ChartBar.vue
