+++ frontend/components/statistics/ChartDoughnut.vue
